+++ JokEngine/JoustExample/PlayerKnight.cpp
@@ -20,7 +20,7 @@ void PlayerKnight::Update()
 {
 	Knight::Update();
 	std::vector<Command*> commands = input->HandleInput();
-	for (int i = 0; i < commands.size(); i++)
+	for (size_t i = 0; i < commands.size(); i++)
 	{
 		commands.at(i)->Execute(*this);
 	}
